CB-226 [refactor]: change product Entity struct
상품 정보를 나타내기 위해서 불필요하게 나누어진 엔티티들을 하나로 모아 사용하기 위해 대대적인
수정을 시작하였습니다.
