+++ src/main/java/com/pinkdumbell/cocobob/domain/product/ProductController.java
@@ -2,6 +2,7 @@ package com.pinkdumbell.cocobob.domain.product;
 
 import com.pinkdumbell.cocobob.common.dto.CommonResponseDto;
 import com.pinkdumbell.cocobob.domain.product.dto.FindAllResponseDto;
+import com.pinkdumbell.cocobob.domain.product.dto.ProductDetailResponseDto;
 import io.swagger.annotations.ApiOperation;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -9,6 +10,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -38,4 +40,5 @@ public class ProductController {
             new ProvideAllResponseClass(HttpStatus.OK.value(), "SUCCESS LOAD PROUDCTS",
                 "상품 가져오기 성공", allProducts));
     }
+
 }
