+++ src/main/java/com/pinkdumbell/cocobob/domain/product/ProductRepository.java
@@ -1,7 +1,8 @@
 package com.pinkdumbell.cocobob.domain.product;
 
+import com.pinkdumbell.cocobob.domain.product.dto.ProductDetailResponseDto;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
-
 }
